+++ src/dev/theskidster/xjge/level/LevelTest.java
@@ -1,7 +1,6 @@
 package dev.theskidster.xjge.level;
 
-import dev.theskidster.xjge.entities.Entity3DAnimTest;
-import dev.theskidster.xjge.entities.EntityBody;
+import dev.theskidster.xjge.entities.EntityBuster;
 import dev.theskidster.xjge.graphics.Light;
 import dev.theskidster.xjge.graphics.Skybox;
 import dev.theskidster.xjge.main.App;
@@ -20,8 +19,7 @@ import org.joml.Vector3f;
  */
 public class LevelTest extends Level {
     
-    public static EntityBody mario = new EntityBody(new Vector3f(0, -10, -40));
-    public static Entity3DAnimTest test = new Entity3DAnimTest(new Vector3f(50, 0, -40));
+    public static EntityBuster buster = new EntityBuster(new Vector3f(0, -10, -40));
     
     @Override
     public void init() {
@@ -31,10 +29,7 @@ public class LevelTest extends Level {
         setSkybox(Skybox.NOON);
         setWorldLight(Light.NOON);
         
-        //test.model.setAnimationSpeed(0.25f);
-        
-        entityList.add(test);
-        entityList.add(mario);
+        entityList.add(buster);
     }
 
     @Override